requirements.txt
# Models
django-model-utils

# Unicode slugification
awesome-slugify

# Images
Pillow

# imagekit for resized images, thumbnails..
django-imagekit

# docs
Sphinx
sphinx-rtd-theme

# package upload
twine

# jwt authentication
djangorestframework-jwt
